# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OPENCLAW OMEGA - Multi-agent orchestrator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: agents
    description: Agent management
  - name: messages
    description: Chat messages
  - name: channels
    description: Channels / workspaces
  - name: tasks
    description: Agent tasks
  - name: telemetry
    description: Agent telemetry and internal state
  - name: swarm
    description: Swarm control
  - name: auth
    description: Operator authentication
  - name: vault
    description: Encrypted secrets vault (operator only)
  - name: social
    description: Official social platform integrations (operator only)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"

  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent by ID
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"

  /agents/{agentId}/telemetry:
    get:
      operationId: getAgentTelemetry
      tags: [telemetry]
      summary: Get agent telemetry and internal state
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Telemetry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentTelemetry"

  /agents/{agentId}/tasks:
    get:
      operationId: listAgentTasks
      tags: [tasks]
      summary: List tasks for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List all channels
      responses:
        "200":
          description: Channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"

  /channels/{channelId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      responses:
        "200":
          description: All tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task status
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /swarm/status:
    get:
      operationId: getSwarmStatus
      tags: [swarm]
      summary: Get aggregate swarm status and stats
      responses:
        "200":
          description: Swarm status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwarmStatus"

  /swarm/pause:
    post:
      operationId: pauseSwarm
      tags: [swarm]
      summary: Pause all agent execution
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwarmStatus"

  /swarm/resume:
    post:
      operationId: resumeSwarm
      tags: [swarm]
      summary: Resume all agent execution
      responses:
        "200":
          description: Resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwarmStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in as an operator (sets an HttpOnly session cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid operator password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out (clears the session cookie)
      responses:
        "200":
          description: Signed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /auth/me:
    get:
      operationId: getAuthStatus
      tags: [auth]
      summary: Whether the caller holds a valid operator session
      responses:
        "200":
          description: Authentication status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /social/platforms:
    get:
      operationId: listSocialPlatforms
      tags: [social]
      summary: List official social platforms, their docs/console URLs, and live connection status
      responses:
        "200":
          description: List of social platform integrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialPlatform"
        "401":
          description: Operator sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /vault:
    get:
      operationId: listVaultSecrets
      tags: [vault]
      summary: List stored secrets (metadata only, values never returned)
      responses:
        "200":
          description: List of secret metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultSecret"
        "401":
          description: Operator sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: setVaultSecret
      tags: [vault]
      summary: Create or update a secret (encrypted at rest)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetVaultSecretRequest"
      responses:
        "200":
          description: Stored secret metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultSecret"
        "400":
          description: Invalid secret data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Operator sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /vault/{name}:
    delete:
      operationId: deleteVaultSecret
      tags: [vault]
      summary: Delete a secret by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteVaultSecretResult"
        "401":
          description: Operator sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Secret not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Agent:
      type: object
      required: [id, name, role, status, color, contextUsed, contextMax, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [idle, thinking, executing, waiting, stalled, hitl]
        color:
          type: string
        avatarInitials:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        contextUsed:
          type: integer
        contextMax:
          type: integer
        capabilities:
          type: array
          items:
            type: string
        createdAt:
          type: string

    AgentInput:
      type: object
      required: [name, role, color]
      properties:
        name:
          type: string
        role:
          type: string
        description:
          type: string
        color:
          type: string
        model:
          type: string
        capabilities:
          type: array
          items:
            type: string

    AgentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [idle, thinking, executing, waiting, stalled, hitl]
        description:
          type: string
        model:
          type: string

    AgentTelemetry:
      type: object
      required: [agentId, monologue, toolCalls, contextUsed, contextMax, updatedAt]
      properties:
        agentId:
          type: integer
        monologue:
          type: array
          items:
            $ref: "#/components/schemas/MonologueLine"
        toolCalls:
          type: array
          items:
            $ref: "#/components/schemas/ToolCall"
        contextUsed:
          type: integer
        contextMax:
          type: integer
        updatedAt:
          type: string

    MonologueLine:
      type: object
      required: [id, text, timestamp, type]
      properties:
        id:
          type: integer
        text:
          type: string
        timestamp:
          type: string
        type:
          type: string
          enum: [thought, action, observation, conclusion]

    ToolCall:
      type: object
      required: [id, toolName, status, startedAt]
      properties:
        id:
          type: integer
        toolName:
          type: string
        args:
          type: ["string", "null"]
        result:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, running, success, error]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]

    Channel:
      type: object
      required: [id, name, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [general, swarm, terminal, logs, hitl]
        description:
          type: ["string", "null"]
        unreadCount:
          type: integer
        lastActivity:
          type: ["string", "null"]
        createdAt:
          type: string

    ChannelInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [general, swarm, terminal, logs, hitl]
        description:
          type: string

    Message:
      type: object
      required: [id, channelId, content, messageType, timestamp]
      properties:
        id:
          type: integer
        channelId:
          type: integer
        agentId:
          type: ["integer", "null"]
        agentName:
          type: ["string", "null"]
        agentColor:
          type: ["string", "null"]
        content:
          type: string
        messageType:
          type: string
          enum: [user, agent, system, tool_output, hitl_request]
        metadata:
          type: ["string", "null"]
        timestamp:
          type: string

    MessageInput:
      type: object
      required: [content, messageType]
      properties:
        content:
          type: string
        messageType:
          type: string
          enum: [user, agent, system, tool_output, hitl_request]
        agentId:
          type: integer
        targetAgentId:
          type: integer

    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        agentId:
          type: ["integer", "null"]
        agentName:
          type: ["string", "null"]
        status:
          type: string
          enum: [queued, running, paused, completed, failed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        progress:
          type: integer
        channelId:
          type: ["integer", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]

    TaskInput:
      type: object
      required: [title, priority]
      properties:
        title:
          type: string
        description:
          type: string
        agentId:
          type: integer
        channelId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, critical]

    TaskUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [queued, running, paused, completed, failed]
        progress:
          type: integer
        agentId:
          type: integer

    SwarmStatus:
      type: object
      required: [paused, activeAgents, totalAgents, runningTasks, completedTasks, totalMessages]
      properties:
        paused:
          type: boolean
        activeAgents:
          type: integer
        totalAgents:
          type: integer
        runningTasks:
          type: integer
        completedTasks:
          type: integer
        totalMessages:
          type: integer
        uptimeSeconds:
          type: integer

    SocialPlatform:
      type: object
      required: [key, displayName, apiBase, docsUrl, consoleUrl, connected]
      properties:
        key:
          type: string
        displayName:
          type: string
        apiBase:
          type: string
        docsUrl:
          type: string
        consoleUrl:
          type: string
        connected:
          type: boolean

    VaultSecret:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    SetVaultSecretRequest:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
        description:
          type: string

    DeleteVaultSecretResult:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: string

    LoginRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AuthStatus:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string